Add a Makefile, and use header files.
The Makefile assumes that the compiler is GCC, or is GCC-like enough (Clang
is) that it least supports the `-MD`, `-MF`, and `-MP` flags.  It uses
these options to automatically track dependencies on header files--this way
when a header file changes, it knows that C files that included it need to
be recompiled.

Each C file (with the exception of the one containing `main()`) first-thing
`#include`s a header file with its own interface.  Including your own
interface isn't strictly-speaking necessary, but it stops one from
declaring a function with different argument or return types between the
signature and the definition, which could otherwise happen!  Including it
as the very-first thing ensures that any header files that the interface
header file depends on are included in the interface header file.

After including the interface header file, each file then includes any
system or user header files.  After each `#include` statement is a comment
noting what that header provides that the C file uses.  I know I got all of
the functions--it's possible that I missed some structure or type
definitions, or constants.

Each header file is wrapped in an #ifndef/#define such that it may be
included multiple times without evaluating the code inside of it multiple
times.  This doesn't so much matter for header files containing only
function signatures, but it does prevent errors for header files containing
struct definitions, type definitions, macros, constants, et cetera.  An
alternative would have been to put `#pragma once` at the top of each header
file; however this is non-standard, an C programmers generally consider it
a C++ism and don't use it in C programs.

I also renamed math.c/math.h to node_math, to avoid conflicting with libc's
`math.h`.